ADKHack/python_agents/requirements.txt
pytest>=7.0.0
pytest-cov>=4.0.0
google-adk>=1.0.0 